requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
